+++ src/main/java/com/kkm/kkm_server_v2/domain/location/domain/repository/LocationRepository.java
@@ -12,4 +12,5 @@ import java.util.Optional;
 public interface LocationRepository extends JpaRepository<Location, Long> {
     List<Location> findAllByUserLocation(User user);
     Optional<Location> findByUserLocationAndAddress(User user, String address);
+    Optional<Location> findByIdAndUserLocation(Long id, User user);
 }
