+++ src/main/java/com/kkm/kkm_server_v2/domain/location/facade/LocationFacade.java
@@ -1,8 +1,10 @@
 package com.kkm.kkm_server_v2.domain.location.facade;
 
+import com.kkm.kkm_server_v2.domain.location.domain.Location;
 import com.kkm.kkm_server_v2.domain.location.domain.repository.LocationRepository;
 import com.kkm.kkm_server_v2.domain.location.exception.LocationAlreadyExistException;
 import com.kkm.kkm_server_v2.domain.location.exception.LocationCountExceedException;
+import com.kkm.kkm_server_v2.domain.location.exception.LocationNotFoundException;
 import com.kkm.kkm_server_v2.domain.user.domain.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
@@ -25,5 +27,10 @@ public class LocationFacade {
         locationRepository.findByUserLocationAndAddress(user, address).ifPresent(location -> {
             throw LocationAlreadyExistException.EXCEPTION;
         });
+    }
+    @Transactional
+    public Location findByIdAndUserLocation(Long id, User user) {
+        return locationRepository.findByIdAndUserLocation(id, user)
+                .orElseThrow(() -> LocationNotFoundException.EXCEPTION);
     }
 }
