+++ src/main/java/com/kkm/kkm_server_v2/domain/user/domain/User.java
@@ -44,15 +44,6 @@ public class User {
     private String imgUrl;
 
     @Column(nullable = false)
-    private double latitude;
-
-    @Column(nullable = false)
-    private double longitude;
-
-    @Column(nullable = false)
-    private String address;
-
-    @Column(nullable = false)
     private int kkm;
 
     @Enumerated(EnumType.STRING)
@@ -107,21 +98,11 @@ public class User {
     @OneToMany(mappedBy = "userLocation", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<Location> locationList;
 
-
-    public void updateAddress(String address, double latitude, double longitude) {
-        this.address = address;
-        this.latitude = latitude;
-        this.longitude = longitude;
-    }
-
     @Builder
-    public User(String nickname, String userId, String imgUrl, double latitude, double longitude, String address, Role role, int kkm) {
+    public User(String nickname, String userId, String imgUrl, Role role, int kkm) {
         this.nickname = nickname;
         this.userId = userId;
         this.imgUrl = imgUrl;
-        this.latitude = latitude;
-        this.longitude = longitude;
-        this.address = address;
         this.role = role;
         this.postList = new ArrayList<>();
         this.kkm = kkm;
