+++ src/main/java/com/kkm/kkm_server_v2/domain/user/presentation/UserController.java
@@ -1,7 +1,6 @@
 package com.kkm.kkm_server_v2.domain.user.presentation;
 
 import com.kkm.kkm_server_v2.domain.user.presentation.dto.request.SignUpRequest;
-import com.kkm.kkm_server_v2.domain.user.presentation.dto.request.UpdateAddressRequest;
 import com.kkm.kkm_server_v2.domain.user.presentation.dto.request.UpdateUserInfoRequest;
 import com.kkm.kkm_server_v2.domain.user.presentation.dto.response.KakaoUserExistResponse;
 import com.kkm.kkm_server_v2.domain.user.presentation.dto.response.MyPageResponse;
@@ -9,18 +8,14 @@ import com.kkm.kkm_server_v2.domain.user.service.CheckUserService;
 import com.kkm.kkm_server_v2.domain.user.service.DivideImageService;
 import com.kkm.kkm_server_v2.domain.user.service.MyPageService;
 import com.kkm.kkm_server_v2.domain.user.service.SignUpService;
-import com.kkm.kkm_server_v2.domain.user.service.UpdateAddressService;
 import com.kkm.kkm_server_v2.domain.user.service.UpdateUserInfoService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.bind.annotation.RestController;
@@ -39,7 +34,6 @@ public class UserController {
     private final CheckUserService checkUserService;
     private final SignUpService signUpService;
     private final UpdateUserInfoService updateUserInfoService;
-    private final UpdateAddressService updateAddressService;
     private final DivideImageService divideImageService;
     private final MyPageService myPageService;
 
@@ -61,11 +55,6 @@ public class UserController {
         updateUserInfoService.execute(request, divideImageService.execute(profileImg));
     }
 
-    @Operation(summary = "주소 업데이트")
-    @PutMapping("/update/address")
-    public void UpdateAddress(@RequestBody @Valid UpdateAddressRequest request) {
-        updateAddressService.execute(request);
-    }
 
     @Operation(summary = "마이페이지")
     @GetMapping("/mypage")
